add a new worker to the cypress e2e tests in order to help with intermittent test failures
refactor the signin_passcode cypress test to split it into 2 test suites, one for active users and one for non active users. Refactor/move the sendEmailAndValidatePasscode function from the signin_passcode test into a shared cypress command so that it can be used by the now split cypress tests
