+++ cypress/integration/ete/sign_in_passcode_active_user.8.cy.ts
@@ -1,116 +1,15 @@
-import { Status } from '../../../src/server/models/okta/User';
 import {
 	randomMailosaurEmail,
 	randomPassword,
 } from '../../support/commands/testUser';
 
-describe('Sign In flow, with passcode', () => {
+describe('Sign In flow, with passcode (part 1)', () => {
 	// set up useful variables
 	const returnUrl =
 		'https://www.theguardian.com/world/2013/jun/09/edward-snowden-nsa-whistleblower-surveillance';
 	const encodedReturnUrl = encodeURIComponent(returnUrl);
 	const appClientId = 'appClientId1';
 	const fromURI = '/oauth2/v1/authorize';
-
-	const sendEmailAndValidatePasscode = ({
-		emailAddress,
-		expectedReturnUrl = 'https://m.code.dev-theguardian.com/',
-		params,
-		expectedEmailBody = 'Your one-time passcode',
-		additionalTests,
-	}: {
-		emailAddress: string;
-		expectedReturnUrl?: string;
-		params?: string;
-		expectedEmailBody?: 'Your one-time passcode' | 'Your verification code';
-		additionalTests?: 'passcode-incorrect' | 'resend-email' | 'change-email';
-	}) => {
-		cy.setCookie('cypress-mock-state', '1'); // passcode send again timer
-		cy.visit(`/signin?${params ? `${params}&` : ''}usePasscodeSignIn=true`);
-		cy.get('input[name=email]').clear().type(emailAddress);
-
-		const timeRequestWasMade = new Date();
-		cy.get('[data-cy="main-form-submit-button"]').click();
-
-		cy.checkForEmailAndGetDetails(emailAddress, timeRequestWasMade).then(
-			({ body, codes }) => {
-				// email
-				expect(body).to.have.string(expectedEmailBody);
-				expect(codes?.length).to.eq(1);
-				const code = codes?.[0].value;
-				expect(code).to.match(/^\d{6}$/);
-
-				// passcode page
-				cy.url().should('include', '/signin/code');
-				cy.contains('Enter your one-time code');
-
-				switch (additionalTests) {
-					case 'resend-email':
-						{
-							const timeRequestWasMade2 = new Date();
-							cy.wait(1000); // wait for the send again button to be enabled
-							cy.contains('send again').click();
-
-							cy.checkForEmailAndGetDetails(
-								emailAddress,
-								timeRequestWasMade2,
-							).then(({ body, codes }) => {
-								// email
-								expect(body).to.have.string(expectedEmailBody);
-								expect(codes?.length).to.eq(1);
-								const code = codes?.[0].value;
-								expect(code).to.match(/^\d{6}$/);
-
-								cy.contains('Sign in');
-								cy.get('input[name=code]').type(code!);
-
-								cy.url().should('include', expectedReturnUrl);
-
-								cy.getTestOktaUser(emailAddress).then((user) => {
-									expect(user.status).to.eq('ACTIVE');
-									expect(user.profile.emailValidated).to.eq(true);
-								});
-							});
-						}
-						break;
-					case 'change-email':
-						cy.contains('try another address').click();
-
-						cy.url().should('include', '/signin');
-						break;
-					case 'passcode-incorrect':
-						cy.contains('Sign in');
-						cy.get('input[name=code]').type(`${+code! + 1}`);
-
-						cy.url().should('include', '/signin/code');
-
-						cy.contains('Incorrect code');
-						cy.get('input[name=code]').clear().type(code!);
-
-						cy.contains('Sign in').click();
-
-						cy.url().should('include', expectedReturnUrl);
-
-						cy.getTestOktaUser(emailAddress).then((user) => {
-							expect(user.status).to.eq('ACTIVE');
-							expect(user.profile.emailValidated).to.eq(true);
-						});
-						break;
-					default: {
-						cy.contains('Sign in');
-						cy.get('input[name=code]').type(code!);
-
-						cy.url().should('include', expectedReturnUrl);
-
-						cy.getTestOktaUser(emailAddress).then((user) => {
-							expect(user.status).to.eq('ACTIVE');
-							expect(user.profile.emailValidated).to.eq(true);
-						});
-					}
-				}
-			},
-		);
-	};
 
 	beforeEach(() => {
 		// Intercept the external redirect pages.
@@ -135,7 +34,7 @@ describe('Sign In flow, with passcode', () => {
 			cy.createTestUser({
 				isUserEmailValidated: true,
 			})?.then(({ emailAddress }) => {
-				sendEmailAndValidatePasscode({
+				cy.sendEmailAndValidatePasscode({
 					emailAddress,
 				});
 			});
@@ -145,7 +44,7 @@ describe('Sign In flow, with passcode', () => {
 			cy.createTestUser({
 				isUserEmailValidated: true,
 			})?.then(({ emailAddress }) => {
-				sendEmailAndValidatePasscode({
+				cy.sendEmailAndValidatePasscode({
 					emailAddress,
 					expectedReturnUrl: returnUrl,
 					params: `returnUrl=${encodedReturnUrl}`,
@@ -157,7 +56,7 @@ describe('Sign In flow, with passcode', () => {
 			cy.createTestUser({
 				isUserEmailValidated: true,
 			})?.then(({ emailAddress }) => {
-				sendEmailAndValidatePasscode({
+				cy.sendEmailAndValidatePasscode({
 					emailAddress,
 					expectedReturnUrl: fromURI,
 					params: `fromURI=${fromURI}&appClientId=${appClientId}`,
@@ -243,7 +142,7 @@ describe('Sign In flow, with passcode', () => {
 			cy.createTestUser({
 				isUserEmailValidated: true,
 			})?.then(({ emailAddress }) => {
-				sendEmailAndValidatePasscode({
+				cy.sendEmailAndValidatePasscode({
 					emailAddress,
 					additionalTests: 'resend-email',
 				});
@@ -254,7 +153,7 @@ describe('Sign In flow, with passcode', () => {
 			cy.createTestUser({
 				isUserEmailValidated: true,
 			})?.then(({ emailAddress }) => {
-				sendEmailAndValidatePasscode({
+				cy.sendEmailAndValidatePasscode({
 					emailAddress,
 					additionalTests: 'change-email',
 				});
@@ -265,7 +164,7 @@ describe('Sign In flow, with passcode', () => {
 			cy.createTestUser({
 				isUserEmailValidated: true,
 			})?.then(({ emailAddress }) => {
-				sendEmailAndValidatePasscode({
+				cy.sendEmailAndValidatePasscode({
 					emailAddress,
 					additionalTests: 'passcode-incorrect',
 				});
@@ -392,130 +291,13 @@ describe('Sign In flow, with passcode', () => {
 						 */
 						cy.visit('/signin?usePasscodeSignIn=true');
 						cy.contains('Sign in with a different email').click();
-						sendEmailAndValidatePasscode({
+						cy.sendEmailAndValidatePasscode({
 							emailAddress,
 							expectedEmailBody: 'Your verification code',
 						});
 					});
 				},
 			);
-		});
-	});
-
-	context('non-ACTIVE user', () => {
-		it('STAGED user - should sign in with passcode', () => {
-			cy.createTestUser({ isGuestUser: true })?.then(({ emailAddress }) => {
-				cy.getTestOktaUser(emailAddress).then((oktaUser) => {
-					expect(oktaUser.status).to.eq(Status.STAGED);
-
-					sendEmailAndValidatePasscode({
-						emailAddress,
-					});
-				});
-			});
-		});
-
-		it('PROVISIONED user - should sign in with passcode', () => {
-			cy.createTestUser({ isGuestUser: true })?.then(({ emailAddress }) => {
-				cy.activateTestOktaUser(emailAddress).then(() => {
-					cy.getTestOktaUser(emailAddress).then((oktaUser) => {
-						expect(oktaUser.status).to.eq(Status.PROVISIONED);
-
-						sendEmailAndValidatePasscode({ emailAddress });
-					});
-				});
-			});
-		});
-
-		it('RECOVERY user - should sign in with passcode', () => {
-			cy.createTestUser({ isGuestUser: false })?.then(({ emailAddress }) => {
-				cy.resetOktaUserPassword(emailAddress).then(() => {
-					cy.getTestOktaUser(emailAddress).then((oktaUser) => {
-						expect(oktaUser.status).to.eq(Status.RECOVERY);
-
-						sendEmailAndValidatePasscode({ emailAddress });
-					});
-				});
-			});
-		});
-
-		it('PASSWORD_EXPIRED user - should sign in with passcode', () => {
-			cy.createTestUser({ isGuestUser: false })?.then(({ emailAddress }) => {
-				cy.expireOktaUserPassword(emailAddress).then(() => {
-					cy.getTestOktaUser(emailAddress).then((oktaUser) => {
-						expect(oktaUser.status).to.eq(Status.PASSWORD_EXPIRED);
-
-						sendEmailAndValidatePasscode({ emailAddress });
-					});
-				});
-			});
-		});
-
-		it('NON_EXISTENT user - should show email sent page with no email sent', () => {
-			const emailAddress = randomMailosaurEmail();
-			cy.visit(`/signin?usePasscodeSignIn=true`);
-
-			cy.contains('Sign in');
-			cy.get('input[name=email]').type(emailAddress);
-			cy.get('[data-cy="main-form-submit-button"]').click();
-
-			// passcode page
-			cy.url().should('include', '/signin/code');
-			cy.contains('Enter your one-time code');
-			cy.contains('Don’t have an account?');
-
-			cy.contains('Sign in');
-			cy.get('input[name=code]').clear().type('123456');
-
-			cy.url().should('include', '/signin/code');
-			cy.contains('Enter your one-time code');
-			cy.contains('Don’t have an account?');
-
-			cy.contains('Incorrect code');
-		});
-
-		it('NON_EXISTENT user - should redirect with error when multiple passcode attempts fail', () => {
-			const emailAddress = randomMailosaurEmail();
-			cy.visit(`/signin?usePasscodeSignIn=true`);
-
-			cy.contains('Sign in');
-			cy.get('input[name=email]').type(emailAddress);
-			cy.get('[data-cy="main-form-submit-button"]').click();
-
-			// passcode page
-			cy.url().should('include', '/signin/code');
-			cy.contains('Enter your one-time code');
-			cy.contains('Don’t have an account?');
-
-			// attempt 1
-			cy.contains('Sign in');
-			cy.get('input[name=code]').type('123456');
-			cy.url().should('include', '/signin/code');
-			cy.contains('Incorrect code');
-
-			// attempt 2
-			cy.get('input[name=code]').type('123456');
-			cy.contains('Sign in').click();
-			cy.url().should('include', '/signin/code');
-			cy.contains('Incorrect code');
-
-			// attempt 3
-			cy.get('input[name=code]').type('123456');
-			cy.contains('Sign in').click();
-			cy.url().should('include', '/signin/code');
-			cy.contains('Incorrect code');
-
-			// attempt 4
-			cy.get('input[name=code]').type('123456');
-			cy.contains('Sign in').click();
-			cy.url().should('include', '/signin/code');
-			cy.contains('Incorrect code');
-
-			// attempt 5
-			cy.get('input[name=code]').type('123456');
-			cy.contains('Sign in').click();
-			cy.url().should('include', '/signin');
-			cy.contains('Your code has expired');
 		});
 	});
 });
